Initial Library Code
This is the initial code for the IIS Express connector. This will start
and stop IIS Express given the correct information on the
instantiation of the class. You will need the full path to the config
file for IIS Express (typically a hidden .vs folder in the root of the
project - drill down until you find the ApplicationHost.config file)
and you will need the site name inside that file that you want to
execute (since there can be more than one site in a file). The
console commands are sent to the List<string> right now.